+++ test/vissen/output/filter/gamma_test.rb
@@ -2,14 +2,10 @@
 
 require 'test_helper'
 
-class TestGridContextTarget
-  include Vissen::Output::GridContext
-end
-
 describe Vissen::Output::Filter::Gamma do
   subject { Vissen::Output::Filter::Gamma }
 
-  let(:context)     { TestGridContextTarget.new 2, 3 }
+  let(:context)     { Vissen::Output::GridContext.new 2, 3 }
   let(:pixel_cloud) { Vissen::Output::PixelCloud.new context }
   let(:filter)      { subject.new context }
   let(:red)         { rand }